+++ Adj/Program/Generation.hs
@@ -1,26 +1,32 @@
+{-# LANGUAGE UndecidableInstances #-}
+
 module Adj.Program.Generation where
 
-import Adj.Auxiliary (Casting (Primary, (=-), (-=)))
-import Adj.Algebra (Category, (.), Functor (map), (-|--), (--|||--), (:*:), (:+:), Flat, Dual, (=-=))
+import Adj.Auxiliary (Casting (Primary, (=-), (-=)), type (=!?=), type (=!!??=), FFGH (FFGH))
+import Adj.Algebra (Category, (.), Functor (map), (-|--), (--|||--), (:*:), (:+:), Identity (Identity), Flat, Dual, (=-=))
 
-newtype Generation p t a = Generation (p a (t (Generation p t a)))
+newtype Generation p f o = Generation
+	((=!!??=) p Identity (f =!?= Generation p f) o)
 
-instance Casting (->) (Generation p t) where
-	type Primary (Generation p t) a = p a (t (Generation p t a))
+instance Casting (->) (Generation p f) where
+	type Primary (Generation p f) o = (=!!??=) p Identity (f =!?= Generation p f) o
 	(=-) (Generation m) = m
 	(-=) m = Generation m
 
-instance
+instance 
 	( Category m
-	, Functor m m t
-	, Casting m (Generation p t)
-	, forall b . Functor m m ((Flat p) b)
-	, forall a . Functor m m ((Dual p) a)
-	, forall b . Casting m (Flat p b)
-	, forall a . Casting m (Dual p a)
-	) => Functor m m (Generation p t) where
-		map m = (=-=) @m @(Generation p t)
-			((-|--) m . (--|||--) m)
+	, Functor m m f
+	, Functor m m Identity
+	, Casting m (Generation p f)
+	, Casting m (f =!?= Generation p f)
+	, Casting m ((=!!??=) p Identity (f =!?= Generation p f))
+	, forall r . Casting m (Flat p (Identity r))
+	, forall l . Casting m (Dual p ((f =!?= Generation p f) l))
+	, forall r . Functor m m ((Flat p) r)
+	, forall l . Functor m m ((Dual p) l)
+	) => Functor m m (Generation p f) where
+	map m = (=-=) (map @m m)
+
 
 type Construction = Generation (:*:)